refactor: update ManagedModel path and use tracker.getSummary()
- Replace manual LDAIMetricSummary construction in ManagedModel.run() with tracker.getSummary()
- Populate resumptionToken in _trackedMetrics on construction so getSummary() returns it
- Accumulate toolCalls in _trackedMetrics via trackToolCall() and trackMetricsOf()
- Update ManagedModel test to mock getSummary(), add evaluator mock, and fix assertions
